requirements.txt
openai>=1.59.0
manim==0.18.1
pydantic==2.10.3
python-dotenv==1.0.1
